Conflicts:
	src/core/CLucene/CLMonolithic.cpp
	src/core/CLucene/StdHeader.cpp
	src/core/CLucene/index/CompoundFile.cpp
	src/core/CLucene/index/DirectoryIndexReader.cpp
	src/core/CLucene/index/DocumentsWriter.cpp
	src/core/CLucene/index/FieldsReader.cpp
	src/core/CLucene/index/IndexFileDeleter.cpp
	src/core/CLucene/index/IndexModifier.cpp
	src/core/CLucene/index/IndexReader.cpp
	src/core/CLucene/index/IndexWriter.h
	src/core/CLucene/index/MultiSegmentReader.cpp
	src/core/CLucene/index/MultipleTermPositions.cpp
	src/core/CLucene/index/MultipleTermPositions.h
	src/core/CLucene/index/SegmentInfos.cpp
	src/core/CLucene/index/SegmentReader.cpp
	src/core/CLucene/index/SegmentTermEnum.cpp
	src/core/CLucene/index/Terms.cpp
	src/core/CLucene/index/Terms.h
	src/core/CLucene/index/_CompoundFile.h
	src/core/CLucene/index/_DocumentsWriter.h
	src/core/CLucene/index/_MultiSegmentReader.h
	src/core/CLucene/index/_SegmentHeader.h
	src/core/CLucene/index/_SegmentTermEnum.h
	src/core/CLucene/index/_Term.h
	src/core/CLucene/queryParser/MultiFieldQueryParser.cpp
	src/core/CLucene/queryParser/QueryParser.cpp
	src/core/CLucene/search/BooleanQuery.cpp
	src/core/CLucene/search/BooleanScorer2.cpp
	src/core/CLucene/search/ConjunctionScorer.cpp
	src/core/CLucene/search/ConstantScoreQuery.cpp
	src/core/CLucene/search/DisjunctionSumScorer.cpp
	src/core/CLucene/search/FilteredTermEnum.cpp
	src/core/CLucene/search/FilteredTermEnum.h
	src/core/CLucene/search/FuzzyQuery.cpp
	src/core/CLucene/search/IndexSearcher.cpp
	src/core/CLucene/search/MatchAllDocsQuery.cpp
	src/core/CLucene/search/MatchAllDocsQuery.h
	src/core/CLucene/search/MultiPhraseQuery.cpp
	src/core/CLucene/search/MultiPhraseQuery.h
	src/core/CLucene/search/PhraseQuery.h
	src/core/CLucene/search/PrefixQuery.cpp
	src/core/CLucene/search/RangeFilter.cpp
	src/core/CLucene/search/RangeQuery.cpp
	src/core/CLucene/search/RangeQuery.h
	src/core/CLucene/search/Similarity.cpp
	src/core/CLucene/search/Similarity.h
	src/core/CLucene/search/TermQuery.cpp
	src/core/CLucene/search/TermQuery.h
	src/core/CLucene/search/_BooleanScorer2.h
	src/core/CLucene/search/_DisjunctionSumScorer.h
	src/core/CLucene/store/FSDirectory.cpp
	src/core/CLucene/store/FSDirectory.h
	src/core/CLucene/store/RAMDirectory.cpp
	src/core/CLucene/store/RAMDirectory.h
	src/core/CMakeLists.txt
	src/test/CMakeLists.txt
	src/test/index/TestIndexReader.cpp
	src/test/index/TestIndexWriter.cpp
	src/test/index/TestReuters.cpp
	src/test/index/TestThreading.cpp
	src/test/queryParser/TestMultiFieldQueryParser.cpp
	src/test/search/TestDateFilter.cpp
	src/test/search/TestQueries.cpp
	src/test/search/TestRangeFilter.cpp
	src/test/search/TestSearch.cpp
	src/test/store/TestStore.cpp

+++ src/core/CLucene/index/IndexFileDeleter.cpp
@@ -2,13 +2,14 @@
 #include <boost/shared_ptr.hpp>
 #include "CLucene/store/Directory.h"
 #include "_IndexFileDeleter.h"
-#include "IndexFileNameFilter.h"
+#include "_IndexFileNameFilter.h"
 #include "Term.h"
 #include "_DocumentsWriter.h"
 #include "_SegmentHeader.h"
 #include "CLucene/LuceneThreads.h"
 #include <algorithm>
 #include <assert.h>
+#include <iostream>
 
 CL_NS_USE(store)
 CL_NS_USE(util)
@@ -92,7 +93,7 @@ void IndexFileDeleter::setInfoStream(std::ostream* infoStream) {
 }
 
 void IndexFileDeleter::message(string message) {
-	(*infoStream) << string("IFD [") << Misc::toString( (int32_t)(_LUCENE_CURRTHREADID) ) << string("]: ") << message << string("\n");
+	(*infoStream) << string("IFD [") << Misc::toString( _LUCENE_CURRTHREADID ) << string("]: ") << message << string("\n");
 }
